requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
mysqlclient>=2.1.1
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
celery>=5.2.7
django-filter>=23.2
drf-yasg>=1.21.5
pytest>=7.3.1
pytest-django>=4.5.2
